refactor: split file_management and image_processing into submodules
Peel cohesive clusters out of two oversized Rust files into sibling
submodules, each using `use super::*` to inherit the parent's imports and
re-exported via `pub use` so the existing command paths in lib.rs are
unchanged:

- file_management/albums.rs    (AlbumItem + album commands, ~260 lines)
- file_management/folder_tree.rs (FolderNode + lazy dir scanning, ~270 lines)
- image_processing/analysis.rs  (histogram/waveform/auto-adjust, ~710 lines)

file_management.rs 3661 -> 3135, image_processing.rs 3262 -> 2555.
