+++ src-tauri/src/file_management.rs
@@ -43,6 +43,11 @@ use crate::mask_generation::MaskDefinition;
 use crate::preset_converter;
 use crate::tagging::COLOR_TAG_PREFIX;
 
+mod albums;
+pub use albums::*;
+mod folder_tree;
+pub use folder_tree::*;
+
 fn resolve_thumbnail_cache_dir(app_handle: &AppHandle) -> std::result::Result<PathBuf, String> {
     let cache_dir = app_handle
         .path()
@@ -511,537 +516,6 @@ pub fn list_images_recursive(
         .collect();
 
     Ok(result_list)
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(tag = "type", rename_all = "camelCase")]
-pub enum AlbumItem {
-    Album {
-        id: String,
-        name: String,
-        icon: Option<String>,
-        images: Vec<String>,
-    },
-    Group {
-        id: String,
-        name: String,
-        icon: Option<String>,
-        children: Vec<AlbumItem>,
-    },
-}
-
-fn get_albums_path(app_handle: &AppHandle) -> Result<PathBuf, String> {
-    let data_dir = app_handle
-        .path()
-        .app_data_dir()
-        .map_err(|e| e.to_string())?;
-    let albums_dir = data_dir.join("albums");
-    if !albums_dir.exists() {
-        fs::create_dir_all(&albums_dir).map_err(|e| e.to_string())?;
-    }
-    Ok(albums_dir.join("albums.json"))
-}
-
-pub fn sort_album_tree(items: &mut [AlbumItem]) {
-    items.sort_by(|a, b| {
-        let get_sort_key = |item: &AlbumItem| match item {
-            AlbumItem::Group { name, .. } => (0, name.to_lowercase()),
-            AlbumItem::Album { name, .. } => (1, name.to_lowercase()),
-        };
-
-        let key_a = get_sort_key(a);
-        let key_b = get_sort_key(b);
-
-        key_a.cmp(&key_b)
-    });
-
-    for item in items.iter_mut() {
-        if let AlbumItem::Group { children, .. } = item {
-            sort_album_tree(children);
-        }
-    }
-}
-
-#[tauri::command]
-pub fn get_albums(app_handle: AppHandle) -> Result<Vec<AlbumItem>, String> {
-    let path = get_albums_path(&app_handle)?;
-    if !path.exists() {
-        return Ok(Vec::new());
-    }
-    let content = fs::read_to_string(path).map_err(|e| e.to_string())?;
-    let mut items: Vec<AlbumItem> = serde_json::from_str(&content).map_err(|e| e.to_string())?;
-    sort_album_tree(&mut items);
-    Ok(items)
-}
-
-#[tauri::command]
-pub fn save_albums(mut tree: Vec<AlbumItem>, app_handle: AppHandle) -> Result<(), String> {
-    let path = get_albums_path(&app_handle)?;
-    sort_album_tree(&mut tree);
-    let json_string = serde_json::to_string_pretty(&tree).map_err(|e| e.to_string())?;
-    fs::write(path, json_string).map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-pub fn add_to_album(
-    album_id: String,
-    paths: Vec<String>,
-    app_handle: AppHandle,
-) -> Result<(), String> {
-    let mut tree = get_albums(app_handle.clone())?;
-
-    fn add_recursive(items: &mut [AlbumItem], target_id: &str, paths_to_add: &Vec<String>) -> bool {
-        for item in items.iter_mut() {
-            #[allow(clippy::collapsible_match)]
-            match item {
-                AlbumItem::Album { id, images, .. } if id == target_id => {
-                    for p in paths_to_add {
-                        if !images.contains(p) {
-                            images.push(p.clone());
-                        }
-                    }
-                    return true;
-                }
-                AlbumItem::Group { children, .. } => {
-                    if add_recursive(children, target_id, paths_to_add) {
-                        return true;
-                    }
-                }
-                _ => {}
-            }
-        }
-        false
-    }
-
-    if add_recursive(&mut tree, &album_id, &paths) {
-        save_albums(tree, app_handle)?;
-    }
-    Ok(())
-}
-
-fn sync_album_path_changes(
-    app_handle: &AppHandle,
-    renames: Option<&HashMap<String, String>>,
-    deletions: Option<&HashSet<String>>,
-    folder_rename: Option<(&str, &str)>,
-) {
-    if let Ok(mut tree) = get_albums(app_handle.clone()) {
-        let mut changed = false;
-
-        fn process_nodes(
-            nodes: &mut [AlbumItem],
-            renames: Option<&HashMap<String, String>>,
-            deletions: Option<&HashSet<String>>,
-            folder_rename: Option<(&str, &str)>,
-            changed: &mut bool,
-        ) {
-            for node in nodes.iter_mut() {
-                match node {
-                    AlbumItem::Album { images, .. } => {
-                        let mut new_images = Vec::new();
-
-                        for img in images.drain(..) {
-                            let mut current_img = img;
-
-                            if let Some((old_folder, new_folder)) = folder_rename {
-                                let img_path = Path::new(&current_img);
-                                let old_path = Path::new(old_folder);
-                                if let Ok(stripped) = img_path.strip_prefix(old_path) {
-                                    let new_img_path = Path::new(new_folder).join(stripped);
-                                    current_img = new_img_path.to_string_lossy().into_owned();
-                                    *changed = true;
-                                }
-                            }
-
-                            if let Some(r) = renames {
-                                if let Some(new_path) = r.get(&current_img) {
-                                    current_img = new_path.clone();
-                                    *changed = true;
-                                } else if let Some((base_path, vc_id)) =
-                                    current_img.rsplit_once("?vc=")
-                                    && let Some(new_base) = r.get(base_path)
-                                {
-                                    current_img = format!("{}?vc={}", new_base, vc_id);
-                                    *changed = true;
-                                }
-                            }
-
-                            let mut is_deleted = false;
-                            if let Some(d) = deletions {
-                                if d.contains(&current_img) {
-                                    is_deleted = true;
-                                } else {
-                                    let img_path = Path::new(&current_img);
-                                    for del_path_str in d {
-                                        let del_path = Path::new(del_path_str);
-                                        if img_path.starts_with(del_path) {
-                                            is_deleted = true;
-                                            break;
-                                        }
-
-                                        if let Some((base_path, _)) =
-                                            current_img.rsplit_once("?vc=")
-                                            && base_path == del_path_str
-                                        {
-                                            is_deleted = true;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-
-                            if !is_deleted {
-                                new_images.push(current_img);
-                            } else {
-                                *changed = true;
-                            }
-                        }
-                        *images = new_images;
-                    }
-                    AlbumItem::Group { children, .. } => {
-                        process_nodes(children, renames, deletions, folder_rename, changed);
-                    }
-                }
-            }
-        }
-
-        process_nodes(&mut tree, renames, deletions, folder_rename, &mut changed);
-
-        if changed {
-            let _ = save_albums(tree, app_handle.clone());
-        }
-    }
-}
-
-#[tauri::command]
-pub fn get_album_images(
-    paths: Vec<String>,
-    app_handle: AppHandle,
-) -> Result<Vec<ImageFile>, String> {
-    let settings = load_settings(app_handle.clone()).unwrap_or_default();
-    let enable_xmp_sync = settings.enable_xmp_sync.unwrap_or(false);
-
-    let result_list: Vec<ImageFile> = paths
-        .into_par_iter()
-        .filter_map(|virtual_path| {
-            let (source_path, sidecar_path) = parse_virtual_path(&virtual_path);
-            if !source_path.exists() {
-                return None;
-            }
-
-            let modified = fs::metadata(&source_path)
-                .ok()
-                .and_then(|m| m.modified().ok())
-                .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-                .map(|d| d.as_secs())
-                .unwrap_or(0);
-
-            let is_virtual_copy = virtual_path.contains("?vc=");
-
-            let (is_edited, tags, rating) = {
-                let mut metadata = crate::exif_processing::load_sidecar(&sidecar_path);
-
-                if enable_xmp_sync
-                    && sync_metadata_from_xmp(&source_path, &mut metadata)
-                    && let Ok(json) = serde_json::to_string_pretty(&metadata)
-                {
-                    let _ = fs::write(&sidecar_path, json);
-                }
-
-                let is_raw = crate::formats::is_raw_file(&source_path);
-                let tm_override =
-                    crate::image_processing::resolve_tonemapper_override(&settings, is_raw);
-                let edited = crate::image_processing::is_image_edited(
-                    &metadata.adjustments,
-                    is_raw,
-                    tm_override,
-                );
-                (edited, metadata.tags, metadata.rating)
-            };
-
-            Some(ImageFile {
-                path: virtual_path,
-                modified,
-                is_edited,
-                tags,
-                exif: None,
-                is_virtual_copy,
-                rating,
-            })
-        })
-        .collect();
-
-    Ok(result_list)
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct FolderNode {
-    pub name: String,
-    pub path: String,
-    pub children: Vec<FolderNode>,
-    pub is_dir: bool,
-    pub image_count: usize,
-    pub has_subdirs: bool,
-    pub modified: u64,
-    pub created: u64,
-}
-
-fn has_subdirs(path: &Path) -> bool {
-    if let Ok(entries) = std::fs::read_dir(path) {
-        for entry in entries.filter_map(Result::ok) {
-            if let Ok(file_type) = entry.file_type()
-                && file_type.is_dir()
-            {
-                let name = entry.file_name();
-                if !name.to_string_lossy().starts_with('.') {
-                    return true;
-                }
-            }
-        }
-    }
-    false
-}
-
-fn scan_dir_lazy(
-    path: &Path,
-    expanded_folders: &HashSet<&str>,
-    show_image_counts: bool,
-    prefetch_one_level: bool,
-) -> Result<(Vec<FolderNode>, usize), std::io::Error> {
-    let mut children_folders = Vec::new();
-    let mut current_dir_image_count = 0;
-
-    let entries = match std::fs::read_dir(path) {
-        Ok(entries) => entries,
-        Err(e) => {
-            log::warn!("Could not scan directory '{}': {}", path.display(), e);
-            return Ok((Vec::new(), 0));
-        }
-    };
-
-    for entry in entries.filter_map(Result::ok) {
-        let current_path = entry.path();
-        let (file_type, modified, created) = match entry.metadata() {
-            Ok(meta) => {
-                let ft = meta.file_type();
-                let mod_time = meta.modified().unwrap_or(std::time::SystemTime::UNIX_EPOCH);
-                let cre_time = meta.created().unwrap_or(mod_time);
-
-                (
-                    ft,
-                    mod_time
-                        .duration_since(std::time::SystemTime::UNIX_EPOCH)
-                        .unwrap_or_default()
-                        .as_secs(),
-                    cre_time
-                        .duration_since(std::time::SystemTime::UNIX_EPOCH)
-                        .unwrap_or_default()
-                        .as_secs(),
-                )
-            }
-            Err(_) => continue,
-        };
-
-        let file_name = entry.file_name();
-        let name_str = file_name.to_string_lossy();
-
-        if name_str.starts_with('.') {
-            continue;
-        }
-
-        if file_type.is_dir() {
-            let path_str = current_path.to_string_lossy().into_owned();
-            let is_expanded = expanded_folders.contains(path_str.as_str());
-
-            let should_scan = is_expanded || prefetch_one_level;
-            let next_prefetch = is_expanded;
-
-            let (grand_children, sub_dir_own_images) = if should_scan {
-                scan_dir_lazy(
-                    &current_path,
-                    expanded_folders,
-                    show_image_counts,
-                    next_prefetch,
-                )?
-            } else {
-                let count = if show_image_counts {
-                    WalkDir::new(&current_path)
-                        .into_iter()
-                        .filter_map(Result::ok)
-                        .filter(|e| {
-                            e.file_type().is_file()
-                                && crate::formats::is_supported_image_file(e.path())
-                        })
-                        .count()
-                } else {
-                    0
-                };
-                (Vec::new(), count)
-            };
-
-            let has_any_subdirs = if should_scan {
-                grand_children.iter().any(|c| c.is_dir)
-            } else {
-                has_subdirs(&current_path)
-            };
-
-            let grand_children_sum: usize = grand_children.iter().map(|c| c.image_count).sum();
-            let total_child_count = sub_dir_own_images + grand_children_sum;
-
-            children_folders.push(FolderNode {
-                name: name_str.into_owned(),
-                path: path_str,
-                children: grand_children,
-                is_dir: true,
-                image_count: total_child_count,
-                has_subdirs: has_any_subdirs,
-                modified,
-                created,
-            });
-        } else if show_image_counts
-            && file_type.is_file()
-            && crate::formats::is_supported_image_file(&current_path)
-        {
-            current_dir_image_count += 1;
-        }
-    }
-
-    children_folders.sort_by_key(|a| a.name.to_lowercase());
-
-    Ok((children_folders, current_dir_image_count))
-}
-
-fn get_folder_tree_sync(
-    path: String,
-    expanded_folders: Vec<String>,
-    show_image_counts: bool,
-) -> Result<FolderNode, String> {
-    let root_path = Path::new(&path);
-    if !root_path.is_dir() {
-        return Err(format!("Directory does not exist: {}", path));
-    }
-
-    let (modified, created) = root_path
-        .metadata()
-        .map(|m| {
-            let mod_time = m.modified().unwrap_or(std::time::SystemTime::UNIX_EPOCH);
-            let cre_time = m.created().unwrap_or(mod_time);
-            (
-                mod_time
-                    .duration_since(std::time::SystemTime::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_secs(),
-                cre_time
-                    .duration_since(std::time::SystemTime::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_secs(),
-            )
-        })
-        .unwrap_or((0, 0));
-
-    let expanded_set: HashSet<&str> = expanded_folders.iter().map(|s| s.as_str()).collect();
-
-    let (children, own_count) = scan_dir_lazy(root_path, &expanded_set, show_image_counts, true)
-        .map_err(|e| e.to_string())?;
-
-    let children_sum: usize = children.iter().map(|c| c.image_count).sum();
-    let has_subdirs = children.iter().any(|c| c.is_dir);
-
-    let name = match root_path.file_name() {
-        Some(n) => n.to_string_lossy().into_owned(),
-        None => {
-            let trimmed = path.trim_end_matches(&['/', '\\'][..]);
-            if trimmed.is_empty() {
-                path.clone()
-            } else {
-                trimmed.to_string()
-            }
-        }
-    };
-
-    Ok(FolderNode {
-        name,
-        path: path.clone(),
-        children,
-        is_dir: true,
-        image_count: own_count + children_sum,
-        has_subdirs,
-        modified,
-        created,
-    })
-}
-
-#[tauri::command]
-pub async fn get_folder_children(
-    path: String,
-    show_image_counts: bool,
-) -> Result<Vec<FolderNode>, String> {
-    match tauri::async_runtime::spawn_blocking(move || {
-        let root_path = Path::new(&path);
-        if !root_path.is_dir() {
-            return Err(format!("Directory does not exist: {}", path));
-        }
-        let empty_set = HashSet::new();
-        let (children, _) = scan_dir_lazy(root_path, &empty_set, show_image_counts, false)
-            .map_err(|e| e.to_string())?;
-
-        Ok(children)
-    })
-    .await
-    {
-        Ok(Ok(children)) => Ok(children),
-        Ok(Err(e)) => Err(e),
-        Err(e) => Err(format!("Task failed: {}", e)),
-    }
-}
-
-#[tauri::command]
-pub async fn get_folder_tree(
-    path: String,
-    expanded_folders: Vec<String>,
-    show_image_counts: bool,
-) -> Result<FolderNode, String> {
-    match tauri::async_runtime::spawn_blocking(move || {
-        get_folder_tree_sync(path, expanded_folders, show_image_counts)
-    })
-    .await
-    {
-        Ok(Ok(folder_node)) => Ok(folder_node),
-        Ok(Err(e)) => Err(e),
-        Err(e) => Err(format!("Failed to execute folder tree task: {}", e)),
-    }
-}
-
-#[tauri::command]
-pub async fn get_pinned_folder_trees(
-    paths: Vec<String>,
-    expanded_folders: Vec<String>,
-    show_image_counts: bool,
-) -> Result<Vec<FolderNode>, String> {
-    let result = tauri::async_runtime::spawn_blocking(move || {
-        let results: Vec<Result<FolderNode, String>> = paths
-            .par_iter()
-            .map(|path| {
-                get_folder_tree_sync(path.clone(), expanded_folders.clone(), show_image_counts)
-            })
-            .collect();
-
-        let mut folder_nodes = Vec::new();
-        for result in results {
-            match result {
-                Ok(node) => folder_nodes.push(node),
-                Err(e) => log::warn!("Failed to get tree for pinned folder: {}", e),
-            }
-        }
-        folder_nodes
-    })
-    .await;
-
-    match result {
-        Ok(nodes) => Ok(nodes),
-        Err(e) => Err(format!("Task failed: {}", e)),
-    }
 }
 
 pub fn read_file_mapped(path: &Path) -> Result<Mmap, ReadFileError> {
