+++ src-tauri/src/file_management/albums.rs
@@ -1,0 +1,262 @@
+use super::*;
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+#[serde(tag = "type", rename_all = "camelCase")]
+pub enum AlbumItem {
+    Album {
+        id: String,
+        name: String,
+        icon: Option<String>,
+        images: Vec<String>,
+    },
+    Group {
+        id: String,
+        name: String,
+        icon: Option<String>,
+        children: Vec<AlbumItem>,
+    },
+}
+
+pub(crate) fn get_albums_path(app_handle: &AppHandle) -> Result<PathBuf, String> {
+    let data_dir = app_handle
+        .path()
+        .app_data_dir()
+        .map_err(|e| e.to_string())?;
+    let albums_dir = data_dir.join("albums");
+    if !albums_dir.exists() {
+        fs::create_dir_all(&albums_dir).map_err(|e| e.to_string())?;
+    }
+    Ok(albums_dir.join("albums.json"))
+}
+
+pub fn sort_album_tree(items: &mut [AlbumItem]) {
+    items.sort_by(|a, b| {
+        let get_sort_key = |item: &AlbumItem| match item {
+            AlbumItem::Group { name, .. } => (0, name.to_lowercase()),
+            AlbumItem::Album { name, .. } => (1, name.to_lowercase()),
+        };
+
+        let key_a = get_sort_key(a);
+        let key_b = get_sort_key(b);
+
+        key_a.cmp(&key_b)
+    });
+
+    for item in items.iter_mut() {
+        if let AlbumItem::Group { children, .. } = item {
+            sort_album_tree(children);
+        }
+    }
+}
+
+#[tauri::command]
+pub fn get_albums(app_handle: AppHandle) -> Result<Vec<AlbumItem>, String> {
+    let path = get_albums_path(&app_handle)?;
+    if !path.exists() {
+        return Ok(Vec::new());
+    }
+    let content = fs::read_to_string(path).map_err(|e| e.to_string())?;
+    let mut items: Vec<AlbumItem> = serde_json::from_str(&content).map_err(|e| e.to_string())?;
+    sort_album_tree(&mut items);
+    Ok(items)
+}
+
+#[tauri::command]
+pub fn save_albums(mut tree: Vec<AlbumItem>, app_handle: AppHandle) -> Result<(), String> {
+    let path = get_albums_path(&app_handle)?;
+    sort_album_tree(&mut tree);
+    let json_string = serde_json::to_string_pretty(&tree).map_err(|e| e.to_string())?;
+    fs::write(path, json_string).map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub fn add_to_album(
+    album_id: String,
+    paths: Vec<String>,
+    app_handle: AppHandle,
+) -> Result<(), String> {
+    let mut tree = get_albums(app_handle.clone())?;
+
+    fn add_recursive(items: &mut [AlbumItem], target_id: &str, paths_to_add: &Vec<String>) -> bool {
+        for item in items.iter_mut() {
+            #[allow(clippy::collapsible_match)]
+            match item {
+                AlbumItem::Album { id, images, .. } if id == target_id => {
+                    for p in paths_to_add {
+                        if !images.contains(p) {
+                            images.push(p.clone());
+                        }
+                    }
+                    return true;
+                }
+                AlbumItem::Group { children, .. } => {
+                    if add_recursive(children, target_id, paths_to_add) {
+                        return true;
+                    }
+                }
+                _ => {}
+            }
+        }
+        false
+    }
+
+    if add_recursive(&mut tree, &album_id, &paths) {
+        save_albums(tree, app_handle)?;
+    }
+    Ok(())
+}
+
+pub(crate) fn sync_album_path_changes(
+    app_handle: &AppHandle,
+    renames: Option<&HashMap<String, String>>,
+    deletions: Option<&HashSet<String>>,
+    folder_rename: Option<(&str, &str)>,
+) {
+    if let Ok(mut tree) = get_albums(app_handle.clone()) {
+        let mut changed = false;
+
+        fn process_nodes(
+            nodes: &mut [AlbumItem],
+            renames: Option<&HashMap<String, String>>,
+            deletions: Option<&HashSet<String>>,
+            folder_rename: Option<(&str, &str)>,
+            changed: &mut bool,
+        ) {
+            for node in nodes.iter_mut() {
+                match node {
+                    AlbumItem::Album { images, .. } => {
+                        let mut new_images = Vec::new();
+
+                        for img in images.drain(..) {
+                            let mut current_img = img;
+
+                            if let Some((old_folder, new_folder)) = folder_rename {
+                                let img_path = Path::new(&current_img);
+                                let old_path = Path::new(old_folder);
+                                if let Ok(stripped) = img_path.strip_prefix(old_path) {
+                                    let new_img_path = Path::new(new_folder).join(stripped);
+                                    current_img = new_img_path.to_string_lossy().into_owned();
+                                    *changed = true;
+                                }
+                            }
+
+                            if let Some(r) = renames {
+                                if let Some(new_path) = r.get(&current_img) {
+                                    current_img = new_path.clone();
+                                    *changed = true;
+                                } else if let Some((base_path, vc_id)) =
+                                    current_img.rsplit_once("?vc=")
+                                    && let Some(new_base) = r.get(base_path)
+                                {
+                                    current_img = format!("{}?vc={}", new_base, vc_id);
+                                    *changed = true;
+                                }
+                            }
+
+                            let mut is_deleted = false;
+                            if let Some(d) = deletions {
+                                if d.contains(&current_img) {
+                                    is_deleted = true;
+                                } else {
+                                    let img_path = Path::new(&current_img);
+                                    for del_path_str in d {
+                                        let del_path = Path::new(del_path_str);
+                                        if img_path.starts_with(del_path) {
+                                            is_deleted = true;
+                                            break;
+                                        }
+
+                                        if let Some((base_path, _)) =
+                                            current_img.rsplit_once("?vc=")
+                                            && base_path == del_path_str
+                                        {
+                                            is_deleted = true;
+                                            break;
+                                        }
+                                    }
+                                }
+                            }
+
+                            if !is_deleted {
+                                new_images.push(current_img);
+                            } else {
+                                *changed = true;
+                            }
+                        }
+                        *images = new_images;
+                    }
+                    AlbumItem::Group { children, .. } => {
+                        process_nodes(children, renames, deletions, folder_rename, changed);
+                    }
+                }
+            }
+        }
+
+        process_nodes(&mut tree, renames, deletions, folder_rename, &mut changed);
+
+        if changed {
+            let _ = save_albums(tree, app_handle.clone());
+        }
+    }
+}
+
+#[tauri::command]
+pub fn get_album_images(
+    paths: Vec<String>,
+    app_handle: AppHandle,
+) -> Result<Vec<ImageFile>, String> {
+    let settings = load_settings(app_handle.clone()).unwrap_or_default();
+    let enable_xmp_sync = settings.enable_xmp_sync.unwrap_or(false);
+
+    let result_list: Vec<ImageFile> = paths
+        .into_par_iter()
+        .filter_map(|virtual_path| {
+            let (source_path, sidecar_path) = parse_virtual_path(&virtual_path);
+            if !source_path.exists() {
+                return None;
+            }
+
+            let modified = fs::metadata(&source_path)
+                .ok()
+                .and_then(|m| m.modified().ok())
+                .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
+                .map(|d| d.as_secs())
+                .unwrap_or(0);
+
+            let is_virtual_copy = virtual_path.contains("?vc=");
+
+            let (is_edited, tags, rating) = {
+                let mut metadata = crate::exif_processing::load_sidecar(&sidecar_path);
+
+                if enable_xmp_sync
+                    && sync_metadata_from_xmp(&source_path, &mut metadata)
+                    && let Ok(json) = serde_json::to_string_pretty(&metadata)
+                {
+                    let _ = fs::write(&sidecar_path, json);
+                }
+
+                let is_raw = crate::formats::is_raw_file(&source_path);
+                let tm_override =
+                    crate::image_processing::resolve_tonemapper_override(&settings, is_raw);
+                let edited = crate::image_processing::is_image_edited(
+                    &metadata.adjustments,
+                    is_raw,
+                    tm_override,
+                );
+                (edited, metadata.tags, metadata.rating)
+            };
+
+            Some(ImageFile {
+                path: virtual_path,
+                modified,
+                is_edited,
+                tags,
+                exif: None,
+                is_virtual_copy,
+                rating,
+            })
+        })
+        .collect();
+
+    Ok(result_list)
+}
