+++ src-tauri/src/file_management/folder_tree.rs
@@ -1,0 +1,271 @@
+use super::*;
+
+#[derive(Serialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct FolderNode {
+    pub name: String,
+    pub path: String,
+    pub children: Vec<FolderNode>,
+    pub is_dir: bool,
+    pub image_count: usize,
+    pub has_subdirs: bool,
+    pub modified: u64,
+    pub created: u64,
+}
+
+fn has_subdirs(path: &Path) -> bool {
+    if let Ok(entries) = std::fs::read_dir(path) {
+        for entry in entries.filter_map(Result::ok) {
+            if let Ok(file_type) = entry.file_type()
+                && file_type.is_dir()
+            {
+                let name = entry.file_name();
+                if !name.to_string_lossy().starts_with('.') {
+                    return true;
+                }
+            }
+        }
+    }
+    false
+}
+
+fn scan_dir_lazy(
+    path: &Path,
+    expanded_folders: &HashSet<&str>,
+    show_image_counts: bool,
+    prefetch_one_level: bool,
+) -> Result<(Vec<FolderNode>, usize), std::io::Error> {
+    let mut children_folders = Vec::new();
+    let mut current_dir_image_count = 0;
+
+    let entries = match std::fs::read_dir(path) {
+        Ok(entries) => entries,
+        Err(e) => {
+            log::warn!("Could not scan directory '{}': {}", path.display(), e);
+            return Ok((Vec::new(), 0));
+        }
+    };
+
+    for entry in entries.filter_map(Result::ok) {
+        let current_path = entry.path();
+        let (file_type, modified, created) = match entry.metadata() {
+            Ok(meta) => {
+                let ft = meta.file_type();
+                let mod_time = meta.modified().unwrap_or(std::time::SystemTime::UNIX_EPOCH);
+                let cre_time = meta.created().unwrap_or(mod_time);
+
+                (
+                    ft,
+                    mod_time
+                        .duration_since(std::time::SystemTime::UNIX_EPOCH)
+                        .unwrap_or_default()
+                        .as_secs(),
+                    cre_time
+                        .duration_since(std::time::SystemTime::UNIX_EPOCH)
+                        .unwrap_or_default()
+                        .as_secs(),
+                )
+            }
+            Err(_) => continue,
+        };
+
+        let file_name = entry.file_name();
+        let name_str = file_name.to_string_lossy();
+
+        if name_str.starts_with('.') {
+            continue;
+        }
+
+        if file_type.is_dir() {
+            let path_str = current_path.to_string_lossy().into_owned();
+            let is_expanded = expanded_folders.contains(path_str.as_str());
+
+            let should_scan = is_expanded || prefetch_one_level;
+            let next_prefetch = is_expanded;
+
+            let (grand_children, sub_dir_own_images) = if should_scan {
+                scan_dir_lazy(
+                    &current_path,
+                    expanded_folders,
+                    show_image_counts,
+                    next_prefetch,
+                )?
+            } else {
+                let count = if show_image_counts {
+                    WalkDir::new(&current_path)
+                        .into_iter()
+                        .filter_map(Result::ok)
+                        .filter(|e| {
+                            e.file_type().is_file()
+                                && crate::formats::is_supported_image_file(e.path())
+                        })
+                        .count()
+                } else {
+                    0
+                };
+                (Vec::new(), count)
+            };
+
+            let has_any_subdirs = if should_scan {
+                grand_children.iter().any(|c| c.is_dir)
+            } else {
+                has_subdirs(&current_path)
+            };
+
+            let grand_children_sum: usize = grand_children.iter().map(|c| c.image_count).sum();
+            let total_child_count = sub_dir_own_images + grand_children_sum;
+
+            children_folders.push(FolderNode {
+                name: name_str.into_owned(),
+                path: path_str,
+                children: grand_children,
+                is_dir: true,
+                image_count: total_child_count,
+                has_subdirs: has_any_subdirs,
+                modified,
+                created,
+            });
+        } else if show_image_counts
+            && file_type.is_file()
+            && crate::formats::is_supported_image_file(&current_path)
+        {
+            current_dir_image_count += 1;
+        }
+    }
+
+    children_folders.sort_by_key(|a| a.name.to_lowercase());
+
+    Ok((children_folders, current_dir_image_count))
+}
+
+fn get_folder_tree_sync(
+    path: String,
+    expanded_folders: Vec<String>,
+    show_image_counts: bool,
+) -> Result<FolderNode, String> {
+    let root_path = Path::new(&path);
+    if !root_path.is_dir() {
+        return Err(format!("Directory does not exist: {}", path));
+    }
+
+    let (modified, created) = root_path
+        .metadata()
+        .map(|m| {
+            let mod_time = m.modified().unwrap_or(std::time::SystemTime::UNIX_EPOCH);
+            let cre_time = m.created().unwrap_or(mod_time);
+            (
+                mod_time
+                    .duration_since(std::time::SystemTime::UNIX_EPOCH)
+                    .unwrap_or_default()
+                    .as_secs(),
+                cre_time
+                    .duration_since(std::time::SystemTime::UNIX_EPOCH)
+                    .unwrap_or_default()
+                    .as_secs(),
+            )
+        })
+        .unwrap_or((0, 0));
+
+    let expanded_set: HashSet<&str> = expanded_folders.iter().map(|s| s.as_str()).collect();
+
+    let (children, own_count) = scan_dir_lazy(root_path, &expanded_set, show_image_counts, true)
+        .map_err(|e| e.to_string())?;
+
+    let children_sum: usize = children.iter().map(|c| c.image_count).sum();
+    let has_subdirs = children.iter().any(|c| c.is_dir);
+
+    let name = match root_path.file_name() {
+        Some(n) => n.to_string_lossy().into_owned(),
+        None => {
+            let trimmed = path.trim_end_matches(&['/', '\\'][..]);
+            if trimmed.is_empty() {
+                path.clone()
+            } else {
+                trimmed.to_string()
+            }
+        }
+    };
+
+    Ok(FolderNode {
+        name,
+        path: path.clone(),
+        children,
+        is_dir: true,
+        image_count: own_count + children_sum,
+        has_subdirs,
+        modified,
+        created,
+    })
+}
+
+#[tauri::command]
+pub async fn get_folder_children(
+    path: String,
+    show_image_counts: bool,
+) -> Result<Vec<FolderNode>, String> {
+    match tauri::async_runtime::spawn_blocking(move || {
+        let root_path = Path::new(&path);
+        if !root_path.is_dir() {
+            return Err(format!("Directory does not exist: {}", path));
+        }
+        let empty_set = HashSet::new();
+        let (children, _) = scan_dir_lazy(root_path, &empty_set, show_image_counts, false)
+            .map_err(|e| e.to_string())?;
+
+        Ok(children)
+    })
+    .await
+    {
+        Ok(Ok(children)) => Ok(children),
+        Ok(Err(e)) => Err(e),
+        Err(e) => Err(format!("Task failed: {}", e)),
+    }
+}
+
+#[tauri::command]
+pub async fn get_folder_tree(
+    path: String,
+    expanded_folders: Vec<String>,
+    show_image_counts: bool,
+) -> Result<FolderNode, String> {
+    match tauri::async_runtime::spawn_blocking(move || {
+        get_folder_tree_sync(path, expanded_folders, show_image_counts)
+    })
+    .await
+    {
+        Ok(Ok(folder_node)) => Ok(folder_node),
+        Ok(Err(e)) => Err(e),
+        Err(e) => Err(format!("Failed to execute folder tree task: {}", e)),
+    }
+}
+
+#[tauri::command]
+pub async fn get_pinned_folder_trees(
+    paths: Vec<String>,
+    expanded_folders: Vec<String>,
+    show_image_counts: bool,
+) -> Result<Vec<FolderNode>, String> {
+    let result = tauri::async_runtime::spawn_blocking(move || {
+        let results: Vec<Result<FolderNode, String>> = paths
+            .par_iter()
+            .map(|path| {
+                get_folder_tree_sync(path.clone(), expanded_folders.clone(), show_image_counts)
+            })
+            .collect();
+
+        let mut folder_nodes = Vec::new();
+        for result in results {
+            match result {
+                Ok(node) => folder_nodes.push(node),
+                Err(e) => log::warn!("Failed to get tree for pinned folder: {}", e),
+            }
+        }
+        folder_nodes
+    })
+    .await;
+
+    match result {
+        Ok(nodes) => Ok(nodes),
+        Err(e) => Err(format!("Task failed: {}", e)),
+    }
+}
